refactor: move the pipestage pipeline and trigger history code to separate module from document context file
