+++ config/config.exs
@@ -99,8 +99,8 @@ config :ex_audit,
     WraftDoc.Document.Instance,
     WraftDoc.Document.Instance.Version,
     WraftDoc.Document.OrganisationField,
-    WraftDoc.Document.Pipeline,
-    WraftDoc.Document.Pipeline.Stage,
+    WraftDoc.Pipelines.Pipeline,
+    WraftDoc.Pipelines.Stages.Stage,
     WraftDoc.Document.Theme,
     WraftDoc.Enterprise.ApprovalSystem,
     WraftDoc.Enterprise.ApprovalSystem,
