+++ lib/wraft_doc/pipelines/stages/stage.ex
@@ -1,4 +1,4 @@
-defmodule WraftDoc.Document.Pipeline.Stage do
+defmodule WraftDoc.Pipelines.Stages.Stage do
   @moduledoc """
   The pipeline stages model.
   """
@@ -7,7 +7,7 @@ defmodule WraftDoc.Document.Pipeline.Stage do
 
   schema "pipe_stage" do
     belongs_to(:content_type, WraftDoc.ContentTypes.ContentType)
-    belongs_to(:pipeline, WraftDoc.Document.Pipeline)
+    belongs_to(:pipeline, WraftDoc.Pipelines.Pipeline)
     belongs_to(:data_template, WraftDoc.DataTemplates.DataTemplate)
     belongs_to(:state, WraftDoc.Enterprise.Flow.State)
     belongs_to(:creator, WraftDoc.Account.User)
