+++ lib/wraft_doc/user/user.ex
@@ -60,7 +60,7 @@ defmodule WraftDoc.Account.User do
       through: [:approval_systems, :instance_approval_systems]
     )
 
-    has_many(:pipelines, WraftDoc.Document.Pipeline, foreign_key: :creator_id)
+    has_many(:pipelines, WraftDoc.Pipelines.Pipeline, foreign_key: :creator_id)
     has_many(:payments, WraftDoc.Enterprise.Membership.Payment, foreign_key: :creator_id)
     has_many(:vendors, WraftDoc.Enterprise.Vendor, foreign_key: :creator_id)
     has_many(:organisation_fields, WraftDoc.Document.OrganisationField, foreign_key: :creator_id)
