+++ test/wraft_doc_web/controllers/pipe_stage_controller_test.exs
@@ -5,7 +5,7 @@ defmodule WraftDocWeb.Api.V1.PipeStageControllerTest do
   use WraftDocWeb.ConnCase
   @moduletag :controller
   import WraftDoc.Factory
-  alias WraftDoc.Document.Pipeline.Stage
+  alias WraftDoc.Pipelines.Stages.Stage
   alias WraftDoc.Repo
 
   describe "create" do
